requirements.txt
# Core Dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2

# Database
PyMySQL==1.1.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.0.5

# Environment
python-dotenv==1.0.0

# Web Forms
Flask-WTF==1.2.1
WTForms==3.0.1
email-validator==2.0.0.post2

# Web Server
gunicorn==21.2.0
itsdangerous==2.1.2
click==8.1.7
python-multipart==0.0.6
vercel==28.19.0
cryptography==41.0.7

# CBIR Dependencies
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.21.0
